add fixes for 64-bit. tests running locally except big-ish ints.

+++ cbits/rshim.h
@@ -6,7 +6,7 @@
 // did this really have to be a macro? BAD MATZ
 unsigned int rtype(VALUE obj);
 VALUE int2fix(int i);
-int fix2int(VALUE x);
+long fix2int(VALUE x);
 double num2dbl(VALUE x);
 unsigned int rb_ary_len(VALUE x);
 VALUE keys(VALUE hash);
